Fix 2.2.3--2.3.0 upgrade script
As reported by github user 1165125080, the two initial "DROP VIEW" commands
were in the wrong order, preventing upgrade from 2.2.3 to 2.3.0.

Note that this extension doesn't store any data in the SQL-level objects, so
preiously users could already DROP the extension in version 2.2.3 and create it
in version 2.3.0 as a workaround for this problem.